Use CameraCompatibilityInfo instead of rotationOverride.
Int (representing enum) rotationOverride contained multiple pieces of information: by how much to rotate in rotate-and-crop, and whether to change sensor orientation, making it difficult to parse and extend.
This refactoring allows sending more information, in a more readable format. It also ensures that no more refactorings of this variable is needed if more data needs to be sent (originally, rotationOverride was a boolean, and transition required the same amount of refactoring).

Flag: EXEMPT refactoring, which would be difficult to guard. No behaviour change.
Test: atest PropertyTest
Test: atest CameraPermissionTest
Test: atest VirtualCameraTest
Test: Manual, tested with and without camera compat mode.
Bug: 432651608
Bug: 365725400
Bug: 414347702
Change-Id: I73bc8ba466fbe09af29cbf5653509cd8cc16c08c