Ignore TimeManagerTest#externalSuggestions test
The TimeManager#suggestExternalTime() API was added for automotive
devices to update the Android time based on an external ECU (e.g. if
there's a separate connectivity module that manages time).

The test is testing this API by using it (Setting a different time and
checking it was set), however the external ECU might be using it at the
same time and reseting the time to the correct one making the test
fail.

Since there's no differentation between external sources it's not
possible to ignore the other sources during testing, so the test is
likely to fail for devices that actually use this API.

Ignore the test in CTS until it is improved to resolve this conflict.

Bug: 323901345
Test: run cts -m CtsTimeTestCases
Change-Id: I37447645b16ba3ad5b965ca8b9580bbf3216af77