Deflake ImeInsetsVisibilityTest
These tests launch a TestActivity with an unfocused EditText,
requestFocus on the EditText and immediately afterwards call
showSoftInput. In most cases the onStartInput for the EditText is
processed before the showSoftInput. However, we are seeing a few rare
failures where the showSoftInput is handled first, and succeeds with the
InputConnection on the DecorView.

With the structure of the test and ImeEventStream, this will wait for
the onStartInput on the EditText, which happens after the showSoftInput.
Then, when waiting for the showSoftInput the test fails, as the event
already took place, and the stream progressed past it.

However, as these tests are not exercising this code path, we're
extracting this behaviour into a separate test for tracking, and
modifying the existing tests to requestFocus during app creation,
similar to the other IME tests.

Flag: EXEMPT testfix
Bug: 433788562
Test: atest ImeInsetsVisibilityTest
  FocusHandlingTest#testRequestFocusCallsOnStartInputBeforeShowSoftInput
Change-Id: If88708222acef4a253d7abe9ebbdff870f4d4ecc