Speculative test improvements for CallStateListenerPermissionTest.
We saw a spate of failures for testRegisterWithNoCallLogPermission
recently.  This test has been quite stable lately.  The only flaw I can
see in this test is it registers a telephony callback and never
unregisters it, so perhaps the per-uid registration limit is being hit on
subsequent re-runs?  Catching registration failures and making sure that
unregistration always happens could help.

Bug: 430090776
Test: This is a test fix (maybe).
Flag: EXEMPT test fix.
Change-Id: I89bada6026a74355141b6f45b9820ef505728a51