CameraITS: Add a test for JPEG_R color consistency.
This change includes a new test added to ensure the consistency between
JCA preview and actual captured JPEG_R image. It skips if the HDR is not
supported on JCA. A preview snapshot is taken and saved to compare to
the actual image captured by JCA.

Next cls will use the unused variable jca_preview_path and also cover
the TODO items added in the cl.

Bug: 427212603
Test: Tested locally on a pixel device
FLAG: TEST_ONLY

Change-Id: I4f70596292430762440897a55d73e20d4f652bc1