fix(magnification cts): fix testA11yNodeInfoVisibility_whenOutOfMagnifiedArea_shouldVisible for Android 14 builds
1. Run the test with magnification fullscreen-only capability
When magnifying and updating the button's visibleToUser property, sometimes the magnification mode switch button is showing (a A11yWindow with type TYPE_MAGNIFICATION_OVERLAY) and confusing the logic, which causes it unexpectedly decides the target button is invisible to users.
The test is not for checking the magnification mode switching behavior but the accessibility API, and the mode switch button is getting in the way and confusing the test. In more modern Android we don't observe this test failing, but in older versions we can avoid that by pre-setting the system magnification capabilities to fullscreen-only so that the mode switch button should not appear when magnifying, which then allows the test to actually exercise the logic it is meant to guard.

2. Call magnification with animated=true
When testing on large-screen devices in landscape mode, setting the magnification scale/center with animated=false can cause the window's transform matrix and position to be calculated incorrectly on older Android builds. This is likely due to a timing issue where the properties are not updated synchronously.
The test is not intended to verify animation behavior. Therefore, as a quick test fix, we now call MagnificationController#setCenter with animated=true. This ensures the window properties update correctly, allowing the test to validate the API logic it is meant to guard.

Bug: 433900209
Flag: EXEMPT bugfix
Test: atest AccessibilityMagnificationTest
Change-Id: I645603e9802c54fa9a3d0ae2fa4037d8be15c251