webkit: move tests that need platform apis.
Move the individual tests that need platform APIs out of the shared
code. There are two relevant kinds of of test here:

- Tests for very old APIs that were removed from the SDK long ago, which
  aren't relevant for the SDK sandbox.
- Tests for printing, which depend on internals of the printing code to
  mock things, and aren't likely to be relevant in the SDK sandbox.

This lets us compile the remainder of the test code with test_current instead
of platform APIs, and enables it to check flags by depending on the relevant
flag static library, instead of getting the framework version that isn't
available at runtime in processes with hidden API enforcement.

Bug: 434770734
Test: atest CtsWebkitTestCases
Test: atest CtsSdkSandboxWebkitTestCases
Flag: TEST_ONLY
Change-Id: I2b082c8f711d2ea66508e8a90133afa159a53809