CTS: max users is for switchable (using Tradefed)
Updates CTS tests to use the newer understanding of max users, in that
we have methods to determine the max number of users of a particular
type, and the max number of switchable users, but no longer have an
overall max number of users (of all types) on the device.

These are CTS tests that rely on Tradefed, and can now rely on the newly
introduced Tradefed methods to properly handle max user numbers.

Note that the flag guarded is already done within Tradefed, so we only
require explicit flag guarding in very few cases.

Test: TestDeviceTest
Test: CtsAccountsHostTestCases CtsAppCloningHostTest CtsAppDataIsolationHostTestCases CtsPackageManagerHostTestCases CtsAppSecurityUtils CtsResumeOnRebootHostTestCases CtsSyncContentHostTestCases CtsDevicePolicyManagerTestCases CtsMediaHostTestCases CtsMultiUserHostTestCases CtsOsHostTestCases CtsPackageManagerMultiUserHostTestCases CtsAppCloningMediaProviderHostTest
Flag: android.multiuser.decouple_max_users_from_profiles
Bug: 394178333
Bug: 407111025
Bug: 413464199
Change-Id: I7cda08890b009d8b16065929c4a30c4401bde55d