Remove verification of 16k-bit RSA APK signing keys
The PkgInstallSignatureVerificationTest verifies that APKs signed
with all supported key sizes and signature algorithms work as
expected on all devices, including verification of APKs signed with
16k-bit RSA keys. In an upcoming update, BoringSSL is removing
support for 16k-bit RSA keys, so this commit updates the tests to
remove verification of APKs signed with these keys. Before removing
support, it has been confirmed that there are no packages that are
signed by keys that would be impacted by this change.

Bug: 434905469
Flag: TEST_ONLY
Test: atest PkgInstallSignatureVerificationTest
Change-Id: Ic6d455e29130a5b1d755a4eec66a82e57b6b8d42