videocodec: do not access private fields using reflection
video encoder roi test is making use of QpOffsetRect class to pass roi
metadata to the component in running state. This class is designed to
configure roi metadata if the roi type is rect. If roi type is map then
the configuration is a byte array. While testing roi type map, the test
is converting QpOffsetRect to byte array. The fields of QpOffsetRect are
private. So reflection is used to access this data and then equivalent
byte array is created. This is causing problems because the test suite
is recently updated to not access hidden / non-public apis. Replacing
the test params from QpOffsetRect to a user-defined data type should
resolve this issue.

Bug: 433473416
Test: atest CtsVideoCodecTestCases:VideoEncoderRoiTest
Flag: TEST_ONLY
Change-Id: I6a89f6ebd436c9b1785d022028f90fdb0a307873