media: do not raise failures on non test threads
MediaCodec callback thread communicates available input, output buffer
indices to the client via onInputBufferAvailable and
onOutputBufferAvailable. The test on receiving output buffer index is
verifying if the decoded output is as expected. This verification logic
is part of onOutputBufferAvailable and is invoked by the callback thread
itself. If the results are not as expected, junit assertions are raised
immediately. This is causing the whole process to crash and interrupting
the test run. This change moves assertions from event handler threads
to the actual test.

Bug: 431781192
Test: atest CtsMediaDecoderTestCases:HdrToSdrDecoderTest
Test: atest MctsMediaDecoderTestCases:HdrToSdrDecoderTest
Test: atest CtsMediaDecoderTestCases:HDRDecoderTest
Test: atest MctsMediaDecoderTestCases:HDRDecoderTest
Flag: TEST_ONLY
Change-Id: Id4a8da4add6e591f630109325f37ece4ec6da070