test(photopicker): Skip tests on unsupported hardware
The Photo Picker CTS tests are not designed for Watches, TVs, Auto, or
embedded IoT devices. Running these UI tests on unsupported form factors
can lead to test failures.

This change introduces a check to ensure the tests are only run on
supported hardware (like phones and tablets). A helper function
`isHardwareSupported()` is added to `PhotoPickerTestRule` and is called
from a `@Before` method in the test classes using `Assume.assumeTrue()`
to skip the tests gracefully.

Bug: b/434901829
Bug: b/434717211
Bug: b/435486461
Test: presubmits
Flag: EXEMPT test only
Change-Id: I80444e9f1fabc81a0c984c74c9a3be333f9606fe