veq: do not fail for non pc class devices
veq test was updated to run on all devices with sdk at least 34. The
results of run are stored in a text file. Devices with media performance
class at least 14 MUST meet the test's requirements. Non performance
class devices need not meet the requirements but indicate their failure
as assumption fail. This part is not covered thoroughly at all places.
This is causing the test to fail on non-mediapc devices.

Bug: 435699098
Test: atest CtsVideoEncodingQualityHostTestCases
Flag: TEST_ONLY
Change-Id: I2ca25c6b0c6dca0f33dd90f9f945171284060af5