Fix invalid testPocCVE_2023_21276
The test for testPocCVE_2023_21276 is invalid in many ways: It twiddles
private members of CursorWindow, leading to impossible and inconsistent
states that cannot be replicated using public APIs. Specifically it
changes the `mAllocOffset` private member to '1', which indicates that
the size of the string data is `1`.  The test also sets the total size
of the CursorWindow to zero, and does not write any string data to the
CursorWindow. This is an impossible state. It also checks memory that is
completely outside the bounds of what is written by
CursorWindow.writeToParcel, which is something that CursorWindow has no
control over.

This test is insidious because if the memory layout of CursorWindow is
changed (e.g. a member is added), this test could fail spuriously. These
spurious failures caused aosp/3308541 to have to be rolled back, and
will continue to cause spurious failures if members of CursorWindow are
modified.

Update the test to use public APIs and check for uninitialized heap data
between the string blob and the slots, which was the intent of the test.

See the attached bug for more technical details.

Bug: 437248474
Test: CVE_2023_21276
Test: verified the POC fails on older devices without the fixes
Flag: EXEMPT bug fix
Change-Id: I53f004571936045c0857e63336756e8064bd69b8