Use metalava as a parser for test-current.txt
Replace the hand-written parser with metalava which already implements
this for the rest of the system.

This CL does not change the public API of the testapiparser package,
types are still strings that need to be parsed by other modules. This
will be addressed in later CLs.

Bug: 436548677
Test: atest android.devicepolicy.cts.FunTest
Flag: TEST_ONLY
Change-Id: Icc33284f2167a4bcf9e8f011ea4dffa96630eddb