Change RoundedCornerTests for desktop setup
This test asserts that rounded corner insets value match the display
rounded corner insets if included and are non-existent if not included.
Therefore it is only expected to have rounded corner insets if the
activity is in fullscreen windowing mode and explicitly requests to
include rounded corner in insets.

This CL fixed the existing test case for fullscreen, as well as added
two test cases for the freeform cases.

This CL replaced ActivityTestRule with ActivityScenario, because only
the latter supports ActivityOptions. Note verifyRoundedCorners()
needs to be run in a non-main thread, so this CL's use of
ActivityScenario doesn't follow the recommendations to only include
activity operations in onActivity().

Bug: 437369884
Bug: 354675563
Test: atest RoundedCornerTests on Ika/Cuttlefish Desktop
Flag: EXEMPT bug fix in test
Change-Id: I45e9ca5cc6de722ed21e7ac8ddbbea5a3008964a