Test restoring IME visibility in profile switching
This adds a test to verify that the IME visibility can be restored when
going from one app in some user profile, to another app (that previously
requested the IME) in another user profile.

This also fixes the IME CTS MainActivity logic for reading Integers from
Intent Extras (introduced in [1]). In instant app mode, the intent is
launched with action Browsable, and the extras are included as URI
parameters. The Integer reading method was directly accessing the
extras, instead of checking whether we're in instant mode, and resolving
them from the URI as needed.

  [1]: I100ea0d085b28dc8abb4191e05778bf3bef59252

Flag: EXEMPT TEST_ONLY
Test: atest MultiUserMockImeTest#testProfileSwitchingCanRestoreImeVisibility
  MultiUserMockImeTest#testHidingKeyboardInSplitScreenWithCrossProfileAppsCanShowKeyboardAgain
Bug: 394783004
Change-Id: Iaa475b5f2dabeb6b8d26627f160b425971d89284