Use try-with instead of try-finally to clean up state
The current problem with try-finally approach is that the exceptions
thrown from the try block get lost if the finally block throws more
exceptions.

Use try-with instead and set up a Resource that will get closed and
clean up the state automatically once the try block finishes the
execution.
Try-with also nicely combines the exceptions from the try block and from
the Resource.close method and no information gets lost.

Refer to the official java documentation to learn more information on how the exceptions are combined.

This only changes devicepolicy.cts.AirplaneModeTest and
devicepolicy.cts.WallpaperTest.
If the pattern turns out to be suitable for the use-case, we can
introduce it to other tests in other CLs.

Bug: 324222685
Test: btest android.devicepolicy.cts.AirplaneModeTest*
Test: btest android.devicepolicy.cts.WallpaperTest*
Flag: TEST_ONLY

Change-Id: I3bedfa5e588c0ae98099d0b2542312c0bdc51f2c