Fix some tests in WindowInsetsPolicyTest for desktop devices
Desktop devices launch new tasks in freeform by default, so assertions
that rely on system decorations (status bar) are invalid if they aren't
launched in fullscreen mode explicitly.

To do that we need to replace ActivityTestRule with ActivityScenario
because only the latter supports ActivityOptions. Note to reduce overly
refactoring to tests this CL doesn't use ActivityScenario in the
recommended way.

This CL also stops waiting for activities to become focused explicity
because ActivityScenario does that behind the scene by waiting for
activities to enter a steady state. They aren't 100% equivalent, but it
should be good enough for us to get rid of the focused wait.

Note WindowInsetsPolicyTest#testForcedConsumedTopInsets has additional
issues that this CL doesn't address. It is possible that that issue is a
system issue instead of a test issue.

Bug: 437369884
Test: atest WindowInsetsPolicyTest on desktop devices.
Flag: EXEMPT bug fix in tests
Change-Id: I4589bf08bafccaee996beff603dce095c30f5dfe