Fix PermissionTest failures on HSUM
PermissionTest fails in HSUM environments because pm and appops commands in the test don't specify a user and are sent to the system user by default, while the test is running for a different user.

This CL fixes the issue by adding --user to pm and appops commands in the test.

Bug: 414099829
Bug: 412808737
Bug: 411554676
Bug: 415412534
Test: atest PermissionTest
Flag: TEST_ONLY
Change-Id: I39417de5aeb3889fd7a1158c8d372b97af5d1571