Skip testOutgoingEmergencyCall test where no emergency capable accounts.
On some wearables, `config_pstnCanPlaceEmergencyCalls` is set in the
Telephony config.xml.  This will mean that there are no Telephony
PhoneAccounts registered with CAPABILITY_PLACE_EMERGENCY_CALLS.  As a
consequence, `testOutgoingEmergencyCall` will fail because Telecom cannot
find a phone account capable of emergency calls.

Adding in an assumption that checks prior to placing the test emergency
call whether there is a phone account on the device that meets the
requirements of the test.

Flag: EXEMPT test only fix.
Bug: 360851785
Test: Re-run test suite.
Change-Id: If7d4580cad18da725be354de78a06e4b51b73388