test: Verify security check for setAccountAttributes
Add a test case to ensure that `setAccountAttributes` can only be
called by the package that authenticates the account.

This change introduces a test that attempts to set attributes on an
account for which the test package is not the authenticator. It
verifies that this operation correctly throws a `SecurityException`,
confirming that the API's security model is properly enforced.

Bug: 423023779
Flag: android.provider.new_account_attributes_api_enabled
Test: atest CtsContactsProviderTestCases:android.provider.cts.contacts.ContactsContract_AccountAttributesTest

	modified:   tests/contactsprovider/src/android/provider/cts/contacts/ContactsContract_AccountAttributesTest.java

Change-Id: I15c37604149be3d9f8d9dae7d361afd9939ebd36