Remove the JobCategoryTest.
The job category APIs were introduced to allow clients to
specify a category for their jobs, with the intention of
influencing scheduling and resource allocation. However,
the plan for this feature has changed, and we are not
moving forward with it at this time.

Bug: 419047126
Bug: 439641599
Test: n/a
Flag: EXEMPT removing android.app.job.job_category_apis
Change-Id: I9061cfbf11bee9e2cdf5ab7280b0a7d3e63d295e