Use OnBackInvokedCallback in ScreenPinningTestActivity
This CL ensures that the ScreenPinningTestActivity behaviour is the same
on Android 15 and older and Android 16 and newer.

This CL disables back navigation in ScreenPinningTestActivity. Since
onBackPressed is not called anymore on Android 16 and above, this CL
migrates the back navigation logic to the new OnBackInvokedCallback
system.

Bug: 437028260
Test: Manual, i.e. verified that back navigation is disabled in
      ScreenPinningTestActivity
Flag: EXEMPT bugfix
Change-Id: I6e30677bef6d3514d6b7d5d19f057b4b5a917b69