Fix CtsWindowManagerDeviceTestCases android.server.wm.WindowUntrustedTouchTest for automotive
The `testWhen...ToastWindow_allowsTouch` tests fail on certain
Automotive targets where the toast does not appear within the test
activity's bounds. This causes both an incorrect bounds assertion
and a missed touch event when the test taps the toast's location.

This change fixes the test by tapping the center of the intersection
of the toast and activity rectangles. If they do not intersect,
the test is skipped.

Bug: 470425951 415961974
Flag: EXEMPT test only fix
Test: Tested on Automotive custom targets
Test: atest CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest#testWhenSelfTextToastWindow_allowsTouch
Merged-In: I2bf346cbaaa2671b575680441943c1d77456086e
Change-Id: Ie70aeb65249f0c600fef3f72946360ec62dd0406