Re-enable SafeBrowsing tests
While there is still an open bug to track the fact that the callback
pattern for safebrowsing "BackToSafety" is not ideal, this CL will
re-enable the tests that were disabled due to being flaky, since they
should no longer be flaky.

This will test the correct behavior for the "proceed" case, as well as
test that the `onSafeBrowsingHit` callback receives the expected block
reason value.

The BackToSafety test case will assert the current implementation
behavior and check that the `WebView.getUrl()` method does report the
correct current URL after the `onPageFinished` callback.
In the future, this test should be updated to assert that the correct
sequence of `onPageStart` and `onPageFinished` callbacks are received,
but that change will likely require a TargetSDK update to implement.

Bug: 40196091
Test: atest CtsWebKitTestCases
Flag: TEST_ONLY

Change-Id: Ib77cc5f155fdffe13c2c08489e0d4acecda7fdfb