Fix potential flake source in WebViewClientTest
The assertions in onPageFinished relied on the queues for onPageStarted,
onLoadResource and onErrorReceived to not have been drained by the time
onPageFinished got invoked. This could potentially happen if a test uses
loadUrl instead of loadUrlAndWaitForCompletion.

This change introduces dedicated state variables to track that the
callbacks are received in the right order.

Bug: 37293648
Test: atest CtsWebkitTestCases
Flag: TEST_ONLY
Change-Id: I28977da9d6cc12d919442fc96b7d835b7b2aa404