Set the scope for setPolicy
Scope is now supported (and required), use it correctly.

The solution is not pretty as scope was implicit in previous APIs. The
current bedstead helpers assume that we can call the same code from all
environments. Calls on the parent are still not supported.

Bug: 434655549
Fixed: 434615264
Flag: android.app.admin.flags.policy_streamlining
Test: Self
Change-Id: I6af5638e58ed085d774be5fa438f32a8a4681570