Fix OutputSwitcherTest in response to ControllerCallback change
The other change in this topic adds an overload that calls (for
backwards compatibility) into the pre-existing method. But
OutputSwitcherTest uses a mock, which disables the default
implementation of the new overload, meaning the pre-existing method
is never called, and this breaks the test.

This change replaces the mock with a Spy, which allows the real
implementation (which calls the pre-existing method) to execute.

Bug: b/396394220
Test: atest CtsMediaRouterTestCases:OutputSwitcherTest
Flag: EXEMPT test only change
Change-Id: I2d4a898ef0ca97aeb0b9db58fd166cdcb5247de0