Don't blcok main thread in MouseToTouchCompatChangeTest
This changes the test set up to call desktopMouseRule.move not from the
activity's main thread but from the instrumentation thread, as "move"
actually blocks the thread to wait for updating the pointer location.

Note that this change is just to make the test healthy and is not to fix
the test flakiness.

Bug: 439445482
Test: MouseToTouchTest
Flag: EXEMPT test fix
Change-Id: I1bb62aa75fd67a8ceb068e78b2fbc3bfd36584dc