Revert^2 "Clean up the Reboot Radio Test"
Clean up the Reboot Radio Test

Remove a bunch of cruft from the reboot radio test.

-Remove unnecessary threads
-Switch to TelephonyCallback
-Abstract the threaded waits into a helper method.


Reason for revert: Working test should not have been reverted;
Cuttlefish exhibits broken and flaky behavior unrelated to test
correctness or validity. It is possible that a 5-second sleep in
cuttlefish's implementation is contributing to the flakiness of
the test on the fake target.

Note: Additional changes in this patch.

-Change the timeouts to be longer to work around possible cuttlefish
 brokenness.
-Change the test name and target method to rebootModem() from
 rebootRadio(). Internally these methods exercise the same
 functionality.

Bug: 438451018
Bug: 440971924
Test: atest TelephonyManagerTest#testRebootRadio ; run on Panther
Change-Id: I877f88aac60f188cf882d0675a0c2a6174454958