[HSUM] Correct directory creation in testRestoreFileAlreadyExists
The test was failing because <file>.mkdirs() was called on the destination image path, creating "image.jpeg" as a directory instead of a file. The subsequent call to MediaProviderTestUtils.stageFile then failed when trying to open this directory for writing. This change removes the incorrect testImageFile1.mkdirs() call and ensures that only the parent directory (baseTestDirectory) is created using mkdirs().

Test: atest MediaStoreTest
Flag: EXEMPT TEST_ONLY
Bug: 441471832
Change-Id: Ica9cf5a121509e469fd9d75c20fcb6a008218247