Adjust WindowMetricsActivityTests bounds to be larger than system min
Window bounds used in WindowMetricsActivityTests are currently smaller
than the system minimum and thus resize to be bigger. However, the test
does not expect this size change. Instead make the test bounds larger so
that they do not need to be adjusted to me meet the system minimum.

Flag: EXEMPT test fix
Test: atest CtsWindowManagerDeviceWindow:WindowMetricsActivityTests
Bug: 438795341
Change-Id: I3a34c67352c1c4ebfbb4ce49abd3071e9af24174