WindowUntrustedTouchTest: wait for animations when hiding bars
TouchHelper waits for animations to finish before injecting events
(using UiAutomation's waitForAnimations logic), but the animations for
hiding system bars aren't covered by that since they're client-side, so
WindowManagerService doesn't know about them. This is resulting in
flakes in WindowUntrustedTouchTest#testWhenTextToastWindow_allowsTouch,
where the taskbar is still visible in front of the toast, causing the
injected touch to open some other app rather than reach the test
activity.

To fix this, use an insets animation callback to wait for the bar hiding
animations to complete before proceeding. This is only done with the
affected test cases because it adds ~1.5s to the duration, so doing it
in the setUp method for all tests would significantly slow down test
runs. The downside is that we'll need to remember to add the
waitForInsetsAnimation() call to any future tests that inject touches
near the edges of the screen.

An alternative would be to set the animation duration scale to 0, but
since some of the test cases deal with animations themselves, that could
cause them to fail or pass incorrectly.

[0]: https://cs.android.com/android/platform/superproject/+/android-latest-release:cts/tests/framework/base/windowmanager/src/android/server/wm/insets/WindowInsetsControllerTests.java;l=1127;drc=2df8423eaca21ff92d2d7f350aed08b5c9a6a94b

Test: $ atest --rerun-until-failure=100 \
            'CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest#testWhenTextToastWindow_allowsTouch'
      (Note that this did not reproduce the flake before, so this is a
       speculative fix.)
Test: $ atest CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
      and compare the run times of the tests to check that they aren't
      being increased by 5s (which would happen if the animation waiting
      was timing out).
Test: $ atest CtsWindowManagerDeviceInsets:android.server.wm.insets.WindowInsetsControllerTests
Bug: 440842960
Flag: TEST_ONLY
Change-Id: Ia73be5cd64321b4af538db2ae1576f2f6894a5c6