Explicitly receive events in VerifyInputEventTest
This will help ensure that the events are not getting sent elsewhere
(maybe to some other activity, if the test runs super fast). This will
also make the error message more explicit if the test fails to receive
an event.

Previously, this could have caused an NPE in the test, which was hard to
read.

Bug: 439999941
Test: atest VerifyInputEventTest
Flag: TEST_ONLY
Change-Id: Id1c4f6d2db1b6ca65619397f1f6fde9a9598f7f2