Allowlist OPEN_EYE_DROPPER action temporarily
This test assures non-public API is exposed from system apps, but we
have to preinstall the application before exposing the public API, so we
have a chicken and egg problem.

Temporarily allowlist android.intent.action.OPEN_EYE_DROPPER so that we
can finish the launch process.

Bug: 441819580
Bug: 399172476
Test: android.signature.cts.intent.IntentTest#shouldNotFindUnexpectedIntents in ABTD
Flag: EXEMPT test fix
Change-Id: I0c0833ebc67f44b67fee142e0e051f61d6f6b385