Fix testBaseApkMissingSignatureAdbInstall.
The test is supposed to install an APK without a v4 signature, but it
was actually installing an APK with a v4 signature. This CL fixed this
so that it now tests what it's supposed to test.

After the fix, the test fails on certain ADB versions because it's known
that those versions don't support incremental install without v4
signature. This CL added an assumption to skip the test on those
versions.

Bug: 441479381
Test: atest CtsIncrementalInstallHostTestCases
Flag: EXEMPT TEST_ONLY

Change-Id: I1a56f28d8d46cf95103b9cdf3170013a4f17a759