Run CTS for ime target behind panel with multiple windowing mode
This change makes ImeInsetsVisibilityTest#
testEditorWontCoveredByImeWhenInputWindowBehindPanel to lunch a test
activity in bott fullscreen and freeform windowing mode. To do so, this
splits the test into two.

We found that the test fails on freeform windowing mode, and it's a real
production issue.
This improves the test coverage by splitting the test into two variants,
one in fullscreen windowing mode, and antoehr in freeform windowing
mode.

Bug: 440265933
Test: ImeInsetsVisibilityTest#testEditorWontCoveredByImeWhenInputWindowBehindPanel_{freeformWindow,fullscreenWindow}
Flag: EXEMPT bug fix
Change-Id: I8fcbe4f44e593c2a6b8eff0eac6cdaa993983240