Fix CtsAngleDumpsysGpuTestApp not building srcs correctly
Move the Android.bp and AndroidManifest.xml to the parent directory
so that the AngleDumpsysGpuTestActivity.java is correctly packaged
in the test app CtsAngleDumpsysGpuTestApp. This should address the
"java.lang.ClassNotFoundException" error. Although this should not
affect the test results of testDumpsysAngleInWhenAngleEnabled and
testDumpsysAngleInWhenAngleDisabled.

Add a log to print the results of `adb shell dumpsys gpu` to help
debugging the test failure.

Flag: TEST_ONLY
Test: atest CtsAngleIntegrationHostTestCases
Bug: b/442933014
Change-Id: Iff82740c350e0dd0bef83f405dce0cbf3caeb009