Refactor ResumeOnRebootHostTest teardown logic when there are multiple users.
This will fix the device to not remain PIN locked when the Resume on Reboot fails.

The `runTestTearDownForUsers` helper method tears down for all users starting with the current active one. This way the `switchUser` method will not throw an exception and it will not bypass the execution.

Bug: 434814918
Flag: EXEMPT TEST_ONLY
Change-Id: Iac88b889f0075473824fa42c60d30304652a8fc2