Reland "Change polling to wait in ServiceWorkerClientTest"
Previously, the two tests used the same origin, which meant that the
service worker registration from the positive test could cause a test
failure in the null service worker test by already being registered,
which causes the `register('sw.js')` call to succeed.

This is fixed by having the two tests use different origins, which
isolates service worker registrations between the tests.

It is likely that the original test hid this issue by checking the
boolean value before it was written.

This reverts commit 582ffa7e4e2936494a7b7e8943fd48dd7e9b6d40.

Fixes: 37293648
Change-Id: I616160456e08110ecb163be46924fe2ec593efc3