Add CTS tests for cross-platform backups and restores
This test verifies that the transport flags are correctly set for
cross-platform backup and restore. It also makes sure that the fields
from the new `#onRestoreFile(FullRestoreDataInput)` method are correctly
set by backing up and restoring a file, and that the new transport
flags, app version code and content version are set on restore.

Bug: 429994661
Bug: 437286518
Bug: 437287293
Bug: 437287318
Bug: 437287411
Bug: 437287501
Bug: 437287635
Bug: 437287744
Bug: 437287814
Bug: 437287864
Bug: 437288024
Test: atest FullBackupLifecycleTest
Flag: TEST_ONLY
Change-Id: I6943ce3c9ce018c4b28bb736a7b5ded21de6099a