Remove the manual WFD tests from CTS-V as they have been replaced by automated ones.
This CL only removes the WFD tests from the CTS-V test list but keeps the source code. We will remove the source code in the next release. This is just in case we receive reports from OEMs and we find that we need to restore these tests.

Flag: TEST_ONLY
Test: http://screenshot/7cj29pcSTeZuDUG
Bug: 443058711
Change-Id: Id5895da89ec3445cf6fb6ac79785106d3169926d