Use multiple tasks flag to launch IntentLaunchActivity
The StrictModeTest cases that verify an UnsafeIntentLaunchViolation
are thrown as expected make use of the IntentLaunchActivity to
launch unparceled Intents in an unsafe manner that would trigger
the violations. However, this test activity is launched with the
new task flag which may reuse an existing activity if one matches
the Intent; this can result in a race condition where the new
activity is restarted before the onStop / onDestroy are called
against it. This commit avoids this race condition by using the
multiple tasks flag which will always start a new activity even
if an existing activity matches the intent.

Fixes: 363965003
Flag: EXEMPT TEST_ONLY
Test: atest --rerun-until-failure 250 StrictModeTest
Change-Id: I39052681d6a28448b06ac2503a7721bd1d1dc0ce