Put default settings result in shared preferences to save result
The settings for bubbles changes when you enable a bubble
for the first time, this causes a cts verifier failure when
not necessary if the user goes through the flow again. This
places the result in shared preferences so it can
be checked if the test fails.

Flag: EXEMPT BUGFIX
Fixes: 419708343
Bug: 418834327
Bug: 421415843
Test: Manual, go through the cts verifier tests, verify default
settings test pass, open a bubble in the next step. Close app and
re-enter bubbles test and verify default settings test still passes

Change-Id: I8b94bb99629c6129dd7b232e7b1d5aa56e2218e1