Fix WindowUntrustedTouchTest for multi-display environments
This change addresses flakiness in `WindowUntrustedTouchTest` when
running on a secondary display for a secondary user. The test was not
fully multi-display compliant because it was hardcoded to use
`Display.DEFAULT_DISPLAY`.

This change replaces the hardcoded default display with the display ID
from the test activity. This ensures the test correctly targets the
activity's display, making it compatible with multi-display
environments.

Test: atest CtsWindowManagerDeviceInput:WindowUntrustedTouchTest Test:
atest --user-type secondary_user_on_secondary_display
CtsWindowManagerDeviceInput:WindowUntrustedTouchTest Bug: 430163384
Flag: TEST_ONLY

Change-Id: I23ce74853cef2b147029b4c333db8472f82c50aa