Resolve auto unhold on call failure CUJ
The auto unhold call failure test was failing due to the connection
being in a dialing state when the call was disconnected. The auto unhold
functionality only happens if the call fails during a setup stage.
Ensuring that the connection fails will ensure that the connection is
not forced into dialing in ManagedConnectionService.

Bug: 445532516
Flag: EXEMPT TEST_ONLY
Test: atest CtsTelecomCujTestCases
Change-Id: I8bd8179ae21d6f92e2fe1e79c8de7c4d6f0a17cb