Cast to ECKey instead of ECPublicKey when checking EC parameters.
The ECKey interface (not ECPublicKey) provides the "getParams()"
method that's needed for checking the EC parameter spec. The
reasoning for why X25519 and Ed25519 public keys can't be cast
to instances of ECKey (or ECPublicKey) is the same: Conscrypt's
public key classes don't implement either interface. Updated the
code comment to explain this more precisely.

Bug: 282063229
Flag: EXEMPT BUGFIX
Test: atest CtsKeyStoreTestCases:android.keystore.cts.KeyAttestationTest
Change-Id: I49cc72db683a72bde6c2617cc7d3b552704dc7fd