Temporarily disable getDmabufTotalExportedKb assertion for android-mainline kernel
The android-mainline kernel does not, and will never set
CONFIG_DMABUF_SYSFS_STATS which is currently necessary to successfully
obtain data for the dmabuf_total_exported_kb field, causing the
SystemMemoryStatsTests#testSystemMemoryAtom test to fail.

There are upcoming changes which remove the need for
CONFIG_DMABUF_SYSFS_STATS for getDmabufTotalExportedKb to work
correctly. Disable this assertion for the android-mainline kernel until
those changes are merged.

The android-mainline kernel version is a moving target, but the latest
release kernel is 6.12.

Bug: 441081044
Flag: EXEMPT TEST_ONLY
Change-Id: I9396247985718239739af12fd0e98c7e3a67ded4