Remove focus check from IME Switcher Menu tests
Currently the IME Switcher Menu tests check that the focused activity is
the TestActivity, but the focused window is not from this activity, as
the IME Switcher Menu is focused, above the TestActivity, but does not
belong to it.

This check breaks if we set config_preventImeStartupUnlessTextEditor, as
there would be no IME parent, which would disallow any windows of type
INPUT_METHOD or INPUT_METHOD_DIALOG from being focusable (logic added
in [1]).

However, the test itself does not depend on the IME Switcher Menu being
focusable, and only used this assertion for simplicitly. This re-works
the assertion to no longer check the focused window, but rather check
that we have a window with the name IME Switcher Menu, that is visible
and has its surface shown. Additionally this adds some calls to wait for
the transition of showing the IME Switcher Menu to finish before
continuing the test.

Further changes to the focusability logic are tracked in 446170524.

  [1]: I3ac91cf02f0270b2d54581aa9174a875836ca692

Flag: EXEMPT TEST_ONLY
Bug: 406501026
Test: atest InputMethodManagerTest#testInputMethodPickerShownItems
  InputMethodManagerTest#testInputMethodPickerSwitchIme
Change-Id: Iaf83ad5b2046dcd53d87c3ac148aa7eab787ea56