Fix failed testEventsFromPreviousSessionIsDiscarded()
This CL reduced the number of back button events from three to two
in the test to navigate back to LoginActivity, as only two are
necessary. The third event was redundant and a source of test
flakiness. The test will inject additional back events (up to 5)
with waits if required in specific scenarios, so this change will
not cause regressions.

Bug: 280116881
Test: atest CtsAutoFillServiceTestCases
Flag: TEST_ONLY
Change-Id: I9aaca4034d3fac06005144ebdf8ee95183b80b8d