Remove obsolete unit test
This test recently becomes flaky because of a change in the CDM Secure Channel. See b/445614723#comment6 for details. Taking a closer look, this test is not necessary because the WSM caller, which is the app that provides the connection, should know if the connection is closed. It does not need to rely on a signal from WSM to know that.

Test: None. Removing a test only
Bug: 445614723
Flag: EXEMPT TEST_ONLY
Change-Id: I0077ecfa5ee13fe98205a0bb5d6cefb5cd6ee06d