[ADI][cts] fix CtsDeveloperVerificationHostTestCases for updated system apps
The test assumed that the system apps involved in this test are not
updated, but in real testing environments they could be updated and
installed in the data partition in a split APK format. Update the test
to accommodate that.

Also fixed an issue where the update owner info can be lost during the
test if the test first resinstalls the verifier via adb.

FLAG: android.content.pm.verification_service
BUG: 447056196
FIXES: 447056196
Test: atest com.android.tests.developerverification.host.DeveloperVerificationHostsideTest

Change-Id: I9dd038d9f8b1873a63670ef2b6fd0296fd7899c1